backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.2
django-grappelli==2.14.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
filelock==3.0.12
idna==2.10
Pillow==8.0.1
pycodestyle==2.7.0
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
virtualenv==20.2.2
